Added a version with working dates (#130)
* Added a version with working dates
* Moved reporter-plywood to a public npm repo, to simplify build process
* Fixed tests
* Added skip tests param, updated k8s config
* Added plywood-server k8s pod

+++ Jenkinsfile
@@ -11,6 +11,7 @@ def kustomizeAndDeploy(overlay, cluster, imageNames) {
             /usr/local/bin/kustomize edit set image ${imageName} && cd -")
     }
     echo "Deploying"
+    sh("kubectl config get-contexts")
     sh("/usr/local/bin/kustomize build kubernetes/overlays/${overlay} | kubectl --context=${cluster} --namespace=${overlay} apply -f -")
 }
 
@@ -20,14 +21,16 @@ node {
 
     properties([
         parameters([
-        booleanParam(name: 'DEPLOY', defaultValue: false, description: 'Deploy this branch to staging')
+            booleanParam(name: 'DEPLOY', defaultValue: false, description: 'Deploy this branch to staging'),
+            booleanParam(name: 'SKIP_TESTS', defaultValue: false, description: 'Skip unit and integration tests')
         ])
     ])
 
     def appName = 'datareporter/datareporter'
     def appNginxName = 'datareporter/nginx'
+    def appPlywoodServerName = 'datareporter/plywood-server'
     def registryRegion = 'eu.gcr.io'
-    def cluster = 'do-fra1-k8s-1-18-8-do-1-fra1'
+    def cluster = 'do-fra1-k8s-1-24-4-do-0-fra1-1665788974923'
     def imageNames = []
 
     sh("git fetch --tags origin")
@@ -56,18 +59,21 @@ node {
 
         stage("Run tests") {
             sh("docker-compose build")
-            lock("tests"){
-                sh("docker-compose up -d postgres")
-                try{
-                    retry(5){
-                        sh("docker-compose run --rm postgres psql -h postgres -U postgres -c \"DROP DATABASE IF EXISTS tests\"")
+            if (params.SKIP_TESTS == false) {
+                lock("tests"){
+                    sh("docker-compose up -d postgres")
+                    try{
+                        retry(5){
+                            sh("docker-compose run --rm postgres psql -h postgres -U postgres -c \"DROP DATABASE IF EXISTS tests\"")
+                        }
+                        sh("docker-compose run --rm postgres psql -h postgres -U postgres -c \"CREATE DATABASE tests\"")
+                        sh("mkdir -p ./logs && touch ./logs/results.xml && chmod 666 ./logs/results.xml")
+                        sh("docker-compose run server tests --junitxml=/app/logs/results.xml")
+                        sh("find . -name *.xml")
+                        junit allowEmptyResults: true, skipPublishingChecks: true, testResults: './logs/results.xml'
+                    } finally {
+                        sh("docker-compose down")
                     }
-                    sh("docker-compose run --rm postgres psql -h postgres -U postgres -c \"CREATE DATABASE tests\"")
-                    sh("docker-compose run server tests --junitxml=/app/result.xml ")
-                    sh "find . -name *.xml"
-                    junit skipPublishingChecks: true, testResults: 'result.xml'
-                }finally{
-                    sh("docker-compose down")
                 }
             }
         }
@@ -82,6 +88,27 @@ node {
             for (tag in imageTags) {
                 echo("Pushing docker image for ${appName} with tag ${tag}")
                 dockerimageDr.push(tag)
+            }
+        }
+
+        stage("Build plywood-server docker image",) {
+            echo "Build docker image for: ${appPlywoodServerName}"
+            def imageNamePlywoodServer = "${registryRegion}/${appPlywoodServerName}:${latestTagRelease}-${shortCommit}"
+
+            dockerimagePlywoodServer = docker.build("${appPlywoodServerName}", "${imageLabel} ${buildArgs} plywood/server")
+            imageNames.add("${registryRegion}/${appPlywoodServerName}=" + imageNamePlywoodServer)
+        }
+
+        stage("Push plywood-server docker image") {
+
+            def imageTags = []
+            imageTags.add("${latestTagRelease}-${shortCommit}")
+            imageTags.add("${shortCommit}_${env.BUILD_ID}")
+            imageTags.add("${env.BRANCH_NAME}".replaceAll("/", "."))
+
+            for (tag in imageTags) {
+                echo("Pushing docker image for ${appPlywoodServerName} with tag ${tag}")
+                dockerimagePlywoodServer.push(tag)
             }
         }
